+++ src/app/config/__init__.py
@@ -1,0 +1,9 @@
+# Local
+from .logger import create_logger
+
+
+LOGGER = create_logger()
+
+__all__ = [
+    "LOGGER",
+]