+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.1] - 2023-08-31
+
+### Added
+
+- Establish .env parameters.
+- First config global variables in config.py.
+- Added database config file.
+- Added logger functionality.
+- Added users SQLAlchemy model.
+- Added pydantic schema users.
+- Added utils module.
+- Added oauth functionality.
+
+### Changed
+
+- docker-compose.yml file modified to more specific purposes.
+
 ## [0.1.0] - 2023-08-31
 
 ### Added